+++ packages/core/src/shared/types.ts
@@ -1,3 +1,5 @@
+import { Key } from "ts-toolbelt/out/Any/Key";
+import { ABConfig } from "..";
 import { FlagValue } from "./schemas";
 
 export enum AbbyEventType {
@@ -23,3 +25,9 @@ export type LegacyAbbyDataResponse = {
   }>;
   flags: Array<{ name: string; isEnabled: boolean }>;
 };
+
+export type ExtractVariants<
+  TestName extends Key,
+  Tests extends Record<TestName, ABConfig>,
+> = Tests[TestName]["variants"][number];
+