+++ packages/react/src/context.tsx
@@ -1,7 +1,6 @@
 import { Abby, AbbyConfig, ABConfig, FlagValueString, FlagValueStringToType } from "@tryabby/core";
 import React, { useCallback, useEffect, useRef, useState, type PropsWithChildren } from "react";
 import { HttpService } from "@tryabby/core";
-import { ABBY_INSTANCE_KEY } from "@tryabby/core";
 import { AbbyDataResponse, AbbyEventType } from "@tryabby/core";
 import { F } from "ts-toolbelt";
 import { FlagStorageService, TestStorageService } from "./StorageService";
@@ -14,12 +13,18 @@ export type withDevtoolsFunction = (
   }
 ) => () => JSX.Element | null;
 
+export type ABTestReturnValue<Lookup, TestVariant> = Lookup extends undefined
+  ? TestVariant
+  : TestVariant extends keyof Lookup
+  ? Lookup[TestVariant]
+  : never;
+
 export function createAbby<
   FlagName extends string,
   TestName extends string,
   Tests extends Record<TestName, ABConfig>,
   Flags extends Record<FlagName, FlagValueString> = Record<FlagName, FlagValueString>,
-  ConfigType extends AbbyConfig<FlagName, Tests> = AbbyConfig<FlagName, Tests>
+  ConfigType extends AbbyConfig<FlagName, Tests> = AbbyConfig<FlagName, Tests>,
 >(abbyConfig: F.Narrow<AbbyConfig<FlagName, Tests, Flags>>) {
   const abby = new Abby<FlagName, TestName, Tests, Flags>(
     abbyConfig,
@@ -64,7 +69,18 @@ export function createAbby<
   // we need to return the config as a const so that the types are narrowed
   const config = abbyConfig as unknown as ConfigType;
 
-  const useAbby = <K extends keyof Tests>(name: K) => {
+  const useAbby = <
+    K extends keyof Tests,
+    TestVariant extends Tests[K]["variants"][number],
+    LookupValue,
+    Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
+  >(
+    name: K,
+    lookupObject?: F.Narrow<Lookup>
+  ): {
+    variant: ABTestReturnValue<Lookup, TestVariant>;
+    onAct: () => void;
+  } => {
     const { tests } = useAbbyData();
 
     // always render an empty string on the first render to avoid SSR mismatches
@@ -76,7 +92,7 @@ export function createAbby<
       const newVariant = tests[name as unknown as TestName]?.selectedVariant;
 
       // should never be undefined after mount
-      if (newVariant != null) {
+      if (newVariant !== undefined) {
         setSelectedVariant(newVariant);
       }
       // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -124,7 +140,11 @@ export function createAbby<
        * It will automatically send the selected variant to the server.
        */
       onAct: onAct,
-      variant: (selectedVariant ?? "") as Tests[K]["variants"][number],
+      variant: lookupObject
+        ? lookupObject[selectedVariant as keyof typeof lookupObject]
+        : // Typescript fails here. If we cast selectedVariant to TestVariant
+          // it still assumes that it is a string. So we cast it to any instead
+          (selectedVariant as any),
     };
   };
 
@@ -168,8 +188,23 @@ export function createAbby<
     return abby.getFeatureFlag(name);
   };
 
-  const getABTestValue = <K extends keyof Tests>(name: K): Tests[K]["variants"][number] => {
-    return abby.getTestVariant(name);
+  const getABTestValue = <
+    TestName extends keyof Tests,
+    TestVariant extends Tests[TestName]["variants"][number],
+    LookupValue,
+    Lookup extends Record<TestVariant, LookupValue> | undefined = undefined,
+  >(
+    testName: TestName,
+    lookupObject?: F.Narrow<Lookup>
+  ): ABTestReturnValue<Lookup, TestVariant> => {
+    const variant = abby.getTestVariant(testName);
+    // Typescript looses its typing here, so we cast as any in favor of having
+    // better type inference for the user
+    if (lookupObject === undefined) {
+      return variant as any;
+    }
+
+    return lookupObject[variant as keyof typeof lookupObject] as any;
   };
 
   const withDevtools: withDevtoolsFunction = (factory, props) => {